+++ gatsby-ssr.js
@@ -1,8 +1,0 @@
-import React from "react"
-
-export const onRenderBody = ({ setHeadComponents, setHtmlAttributes }) => {
-  setHeadComponents([
-    <script data-no-cookie async src="https://cdn.splitbee.io/sb.js" />,
-  ])
-  setHtmlAttributes({ lang: `en` })
-}